refactor(modules): move GoDirectory to module, clean up tests
