+++ src/server/inversify.config.ts
@@ -47,8 +47,8 @@ import { AuthService } from './services/AuthService'
 import { LogoutController } from './controllers/LogoutController'
 import { UrlManagementService } from './modules/user/services'
 import { UserController } from './modules/user'
-import { DirectoryController } from './controllers/DirectoryController'
-import { DirectorySearchService } from './services/DirectorySearchService'
+import { DirectoryController } from './modules/directory'
+import { DirectorySearchService } from './modules/directory/services'
 import { LinkStatisticsController } from './controllers/LinkStatisticsController'
 import { LinkStatisticsService } from './services/LinkStatisticsService'
 import { LinkStatisticsRepository } from './repositories/LinkStatisticsRepository'
