+++ src/server/modules/directory/services/index.ts
@@ -1,0 +1,4 @@
+export {
+  DirectorySearchService,
+  DirectorySearchService as default,
+} from './DirectorySearchService'
